refactor: remove nested `components` folder
