+++ src/components/common/CruiseHeader/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { Layout } from 'antd';
-import logo from '../../assets/logo.png';
+import logo from '../../../assets/logo.png';
 
 const { Header } = Layout;
