+++ src/pages/Agent/AgentTypeBox/index.tsx
